Hoan thanh cac chuc nang chinh

+++ src/views/chinhanh/index.jsx
@@ -6,6 +6,7 @@ import FormDialog from './FormDialog';
 import { useState, useEffect } from 'react';
 import axios from "axios";
 import { useSelector } from 'react-redux';
+import { toast } from 'react-toastify';
 const ChiNhanh = () => {
     const theme = useTheme();
     const colors = tokens(theme.palette.mode);
@@ -22,23 +23,13 @@ const ChiNhanh = () => {
           renderCell: (cellValues) => (
             <Box display="flex" gap={2}>
               <Button
-                sx={{
-                  backgroundColor: colors.greenAccent[600],
-                  '&:hover': {
-                    backgroundColor: colors.greenAccent[500],
-                  },
-                }}
-              >
-                Cập nhật
-              </Button>
-              <Button
               sx={{
                 backgroundColor: colors.redAccent[600],
                 '&:hover': {
                   backgroundColor: colors.redAccent[500],
                 },
               }}
-              onClick={() => handleDeleteChiNhanh(cellValues.row._id)}
+              onClick={() => handleDeleteChiNhanh(cellValues.row._id, cellValues.row.tenChiNhanh)}
               >Xóa</Button>
             </Box>
           ),
@@ -58,21 +49,24 @@ const ChiNhanh = () => {
             // })
             const data = response.data.result;
             setData(data);
+            //console.log(data);
         }
     }).catch((err) => console.log(err))
   };
-  const handleDeleteChiNhanh = async (id) => {
-        axios.delete("http://localhost:3000/api/chinhanhs/"+id,  {
+  const handleDeleteChiNhanh = async (id, tenChiNhanh) => {
+        if(window.confirm(`Bạn có chắc muốn xóa chi nhánh ${tenChiNhanh}?`)){
+          axios.delete("http://localhost:3000/api/chinhanhs/"+id,  {
           headers: {
             Authorization: `Bearer ${token}`
           }
         }).then((res) => {
             if(res.status === 200){
-                alert("Xóa chi nhánh thành công");
+                toast.success("Xóa chi nhánh thành công");
                 setData(data.filter((dt) => dt._id !== id))
                 
             }
         }).catch((err) => console.log(err))
+        }
   };
 
   useEffect(() => {
